Add JSON import in backup options
It is now possible to import JSON data that was exported before. An
import will overwrite existing entries with same uniqueid, and create
new entries for non-existing uniqueid. This commit also adds a dropdown
to select between items, categories and accounts for export data.
CSV export is still not working, because multideviceDatabase lacks some
functions which are included in localstoragedb.